bots/tg/bot4_survey/requirements4.txt
aiogram==3.3.0
gspread==5.11.0
google-auth==2.23.0
python-dotenv==1.0.0
redis>=5.0.0